% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condlike_rep.R
\name{condlike_rep}
\alias{condlike_rep}
\title{Conditional likelihood method for Winner's Curse adjustment using discovery and replication summary statistics - Zhong and Prentice (2008)}
\usage{
condlike_rep(summary_disc, summary_rep, alpha)
}
\arguments{
\item{summary_disc}{Data frame containing summary data from discovery GWAS, three columns: rsid, beta, se}

\item{summary_rep}{Data frame containing summary data from replication GWAS, three columns: rsid, beta, se}

\item{alpha}{significance threshold used in discovery GWAS}
}
\value{
Data frame with only SNPs deemed significance in discovery GWAS, contains their beta and se values from both discovery and replication GWAS
as well as the combined unadjusted estimate, the conditional likelihood MLE and the weighted average for each SNP
}
\description{
Conditional likelihood method for Winner's Curse adjustment using discovery and replication summary statistics - Zhong and Prentice (2008)
}
