% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_bayes.R
\name{empirical_bayes}
\alias{empirical_bayes}
\title{Empirical Bayes correction method for Winner's Curse - simplest implementation - Ferguson et al. (2013)}
\usage{
empirical_bayes(summary_data)
}
\arguments{
\item{summary_data}{Data frame containing summary data, three columns: rsid, beta, se}
}
\value{
Data frame with summary data together with adjusted estimates but (1) for now
}
\description{
Empirical Bayes correction method for Winner's Curse - simplest implementation - Ferguson et al. (2013)
}
