% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR_IQT.R
\name{FDR_IQT}
\alias{FDR_IQT}
\title{FDR Inverse Quantile Transformation method to correct for Winner’s Curse - Bigdeli et al. (2016)}
\usage{
FDR_IQT(summary_data, min_pval = 1e-15)
}
\arguments{
\item{summary_data}{Data frame containing summary data, three columns: rsid, beta, se}

\item{min_pval}{avoids zero p-values (in R) which gives trouble with qnorm(), also abs(z) > 8 is generally not biased}
}
\value{
Data frame with summary data together with adjusted estimates
}
\description{
FDR Inverse Quantile Transformation method to correct for Winner’s Curse - Bigdeli et al. (2016)
}
