% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_stats.R
\name{sim_stats}
\alias{sim_stats}
\title{Simulating GWAS summary statistics}
\usage{
sim_stats(
  nsnp = 10^6,
  h2 = 0.4,
  prop_effect = 0.01,
  nid = 50000,
  rep = FALSE,
  rep_nid = 50000
)
}
\arguments{
\item{nsnp}{A numerical value which specifies the total number of SNPs that
the user wishes to simulate summary statistics for. The default is 1,000,000 SNPs,
i.e. \code{nsnp=10^6}.}

\item{h2}{A numerical value between 0 and 1 which represents the desired
heritability of the trait of interest, or in other words, the total
variance explained in the trait by all SNPs. The default is a moderate
heritability value of 0.4, \code{h2=0.4}.}

\item{prop_effect}{A numerical value between 0 and 1 which determines the
trait's polygenicity, the fraction of the total number of SNPs which are truly associated with the
trait. The default setting is \code{prop_effect = 0.01}.}

\item{nid}{A numerical value which specifies the number of individuals that
the discovery GWAS has been performed with. This value defaults to 50,000 individuals, \code{nid=50000}.}

\item{rep}{A logical value which allows the user to state whether they would
also like to simulate summary statistics for a replication GWAS based on
the same parameters and true effect sizes. The default setting is
\code{rep=FALSE}.}

\item{rep_nid}{A numerical value which specifies the number of individuals
that the replication GWAS has been performed with. Similar to \code{nid},
this value defaults to 50,000 individuals, \code{nid=50000}.}
}
\value{
A list containing three different components in the form of data
  frames, \code{true}, \code{disc} and \code{rep}. The first element,
  \code{true} has two columns, \code{rsid} which contains identification numbers
  for each SNP and \code{true_beta} which is each SNP's simulated true
  association value. \code{disc} has three columns representing the summary statistics
  one would obtain in a discovery GWAS. For each SNP, this data frame contains
  its ID number, its estimated effect size, \code{beta}, and associated standard error, \code{se}.
 Similarly, if the \code{rep} argument in the function has been set to \code{TRUE},
 then the data frame, \code{rep} has three columns representing the summary statistics
  one would obtain in a replication GWAS. In this data frame, just as with \code{disc},
  the values for \code{beta} have been simulated using the true association values, \code{true_beta},
  and the standard errors are reflective of the chosen sample size.
  If \code{rep=FALSE}, \code{NULL} is merely returned for this third element.
}
\description{
\code{sim_stats} is a function which can be used to simulate summary
statistics for a set of independent SNPs for both discovery and replication
GWASs. This function allows the user to create toy datasets with which they
can explore the implementation of the Winner's Curse correction methods.
}
