% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMVCUE.R
\name{UMVCUE}
\alias{UMVCUE}
\title{Winner's Curse adjustment method using discovery and replication summary statistics - Bowden and Dudbridge (2009)}
\usage{
UMVCUE(summary_disc, summary_rep, alpha)
}
\arguments{
\item{summary_disc}{Data frame containing summary data from discovery GWAS, three columns: rsid, beta, se}

\item{summary_rep}{Data frame containing summary data from replication GWAS, three columns: rsid, beta, se}

\item{alpha}{significance threshold used in discovery GWAS}
}
\value{
Data frame with only SNPs deemed significance in discovery GWAS, contains their beta and se values from both discovery and replication GWAS
as well as the UMVCUE for each SNP
}
\description{
Winner's Curse adjustment method using discovery and replication summary statistics - Bowden and Dudbridge (2009)
}
