% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winnerscurse-package.R
\docType{package}
\name{winnerscurse-package}
\alias{winnerscurse-package}
\alias{winnerscurse}
\title{winnerscurse: Winner's Curse Adjustment Methods for Summary Statistics from Genome-Wide Association Studies}
\description{
A package designed to provide users with easy access to published methods which aim to
correct for Winner's Curse using only summary statistics from genome-wide association studies. With merely estimates of effect size and
associated standard error for each genetic variant, users are able to implement these methods to obtain
more accurate estimates of the true effect sizes. These methods can be applied to data from both
quantitative and binary traits.
}
\details{
\strong{Full documentation available here:} \href{https://amandaforde.github.io/winnerscurse/}{https://amandaforde.github.io/winnerscurse/}
}
