% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BR_ss.R
\name{BR_ss}
\alias{BR_ss}
\title{Adaptation of BR-squared method - Faye *et al.* (2011) - suitable for summary statistics}
\usage{
BR_ss(summary_data)
}
\arguments{
\item{summary_data}{Data frame containing summary data, three columns: rsid, beta, se}
}
\value{
Data frame with summary data together with adjusted estimates
}
\description{
Adaptation of BR-squared method - Faye *et al.* (2011) - suitable for summary statistics
}
