% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_interval.R
\name{cl_interval}
\alias{cl_interval}
\title{Confidence interval for conditional likelihood methods - Ghosh et al. (2008)}
\usage{
cl_interval(summary_data, alpha, conf = 0.95)
}
\arguments{
\item{summary_data}{Data frame containing summary data, three columns: rsid, beta, se}

\item{alpha}{significance threshold used in discovery GWAS}

\item{conf}{specifies confidence level, takes a value between 0 and 1 - default is 0.95 corresponding to 95 per cent confidence interval}
}
\value{
Data frame which combines output from application of conditional likelihood methods, i.e. summary data of only significant SNPs together with three corrected estimates, with an upper and a lower bound on the required confidence interval for each SNP
}
\description{
Confidence interval for conditional likelihood methods - Ghosh et al. (2008)
}
