% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_likelihood.R
\name{conditional_likelihood}
\alias{conditional_likelihood}
\title{Conditional likelihood methods to correct for Winner’s Curse - Ghosh et al. (2008)}
\usage{
conditional_likelihood(summary_data, alpha)
}
\arguments{
\item{summary_data}{Data frame containing summary data of only significant SNPs, three columns: rsid, beta, se}

\item{alpha}{significance threshold used in discovery GWAS}
}
\value{
Data frame with summary data together with three corrected estimates
}
\description{
Conditional likelihood methods to correct for Winner’s Curse - Ghosh et al. (2008)
}
