% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_minimizer.R
\name{MSE_minimizer}
\alias{MSE_minimizer}
\title{Approach to combine association estimates from discovery and replication data sets}
\usage{
MSE_minimizer(summary_disc, summary_rep, alpha, spline = FALSE)
}
\arguments{
\item{summary_disc}{Data frame containing summary data from discovery GWAS, three columns: rsid, beta, se}

\item{summary_rep}{Data frame containing summary data from replication GWAS, three columns: rsid, beta, se}

\item{spline}{Logical parameter, default is spline=FALSE, spline=TRUE uses a cubic smoothing spline to predict value of B=summary_disc$beta - summary_rep$beta,
spline=FALSE just uses this raw value for B}
}
\value{
Data frame with only SNPs deemed significance in discovery GWAS, contains their beta and se values from both discovery and replication GWAS
as well as the combined estimate, beta_joint
}
\description{
Approach to combine association estimates from discovery and replication data sets
}
