% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_adjust.R
\name{se_adjust}
\alias{se_adjust}
\title{Obtaining standard errors of adjusted methods via parametric bootstrap}
\usage{
se_adjust(summary_data, method, n_boot = 100)
}
\arguments{
\item{summary_data}{Data frame containing summary data, three columns: rsid, beta, se}

\item{method}{User must specify "BR_ss", "empirical_bayes" or "FDR_IQT"}

\item{n_boot}{number of bootstrap repetitions used - defaults to 100}
}
\value{
Data frame with summary data together with adjusted estimates
}
\description{
Obtaining standard errors of adjusted methods via parametric bootstrap
}
