% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_adjust.R
\name{se_adjust}
\alias{se_adjust}
\title{Standard errors of adjusted discovery GWAS estimates via parametric bootstrap}
\usage{
se_adjust(summary_data, method, n_boot = 100)
}
\arguments{
\item{summary_data}{A data frame containing summary statistics from the
discovery GWAS. It must have three columns with column names \code{rsid},
\code{beta} and \code{se}, respectively, and all columns must contain
numerical values. Each row must correspond to a unique SNP,
identified by the numerical value \code{rsid}.}

\item{method}{A string specifying the function to be implemented on each of
the bootstrap samples. It should take the form \code{"BR_ss"},
\code{"empirical_bayes"} or \code{"FDR_IQT"}.}

\item{n_boot}{A numerical value which determines the number of bootstrap
repetitions to be used. it must be greater than 1. The default value is \code{100}.}
}
\value{
A data frame which combines the output of the chosen method with an
  additional column, namely \code{adj_se}. This column provides the standard
  errors of the adjusted association estimates for each SNP.
}
\description{
\code{se_adjust} is a function which allows the user to obtain approximate
standard errors of adjusted association estimates, by means of parametric
bootstrapping. Standard errors can be evaluated for estimates which have been
corrected with the Empirical Bayes method, FDR Inverse Quantile
Transformation method or the bootstrap method. Note that in comparison to the
other functions in this package, this function can be computationally
intensive and take a several minutes to run, depending on the size of the
data set, the method and the number of bootstraps chosen.
}
\seealso{
\code{\link{empirical_bayes}}, \code{\link{BR_ss}} and
  \code{\link{FDR_IQT}} for details on operation of these methods with
  summary statistics from discovery GWAS.

  \url{https://amandaforde.github.io/winnerscurse/articles/standard_errors_confidence_intervals.html}
   for illustration of the use of \code{se_adjust} with a toy data set and
  further information regarding the manner in which the standard errors are
  computed.
}
